Separate out some index stuff into sidebar.tsx
